+++ src/main/java/org/fundacionjala/coding/marcos/FizzBuzz.java
@@ -1,4 +1,4 @@
-package org.fundacionjala.coding.marcoslara;
+package org.fundacionjala.coding.marcos;
 
 /**
  * Created by Marcos on 8/25/2017.
